requirements.txt
flask
sqlalchemy
pymysql
pillow
pytesseract
PyMuPDF
python-dotenv